[HeapProtector] Add basic infrastructure for Heap Protector
Allow LLVM to generate HeapProtector instrumentation.
HeapProtector has some similarities to ASAN, but a different focus:
Its goal is to provide deterministic, race-free protection against
use-after-reallocation accesses to heap memory, even in multithreaded
environments.

Similarly to HWASAN on non-AArch64 environments, before a pointer can be
accessed, HeapProtector has to both ensure that the access can not cause
use-after-reallocation and decode the provided fat pointer (which is not
a valid native pointer) into a native pointer. However, unlike HWASAN,
the checking and decoding is implemented by calling a function

    __attribute__((preserve_most)) void *__khp_decode_ptr(void *raw_ptr)

which is implemented outside LLVM.

HeapProtector attempts to call that decoding helper as rarely as
possible. For example, given C code like the following:

    int foo(struct bar *bar) {
      int res = bar->a;
      res += bar->b;
      some_other_function(bar);
      return res + bar->c;
    }

HeapProtector should generate code like this:

    int foo(struct bar *bar) {
      struct bar *bar_decoded = __khp_decode_ptr(bar);
      int res = bar_decoded->a;
      res += bar_decoded->b;
      some_other_function(bar);
      return res + bar_decoded->c;
    }

This example demonstrates that an unbounded amount of time can pass
between the __khp_decode_ptr() invocation and the last access through
the decoded pointer.

This patch does not provide any way for the runtime environment to
ensure that the accessed object is not freed and reallocated between
__khp_decode_ptr() and the actual access; such a mechanism will be added
by a later patch.

This patch includes code originally written by Mark Brand.