[HeapProtector] Provide object pin arrays in stack frames
To allow the __khp_decode_ptr() helper to guard against a referenced
object being freed and reallocated after the object liveness check, the
runtime environment must be able to determine which objects are live on
a given stack.

For this purpose, embed an array of object "pins" in each stack frame.
When calling __khp_decode_ptr(), provide a pointer to a slot in that
array as the second argument, and guarantee that the slot will not be
modified afterwards until the decoded pointer is no longer live.

For now, to allow the runtime environment to locate the pin arrays, the
pin arrays form a singly-linked list through the stack, with the list
head stored in a variable in a per-cpu/per-thread variable.

A possible optimization for the future would be to replace the linked
list scheme with proper stack unwinding; however, for the initial
implementation, that would be too much complexity.