[HeapProtector] Add __khp_unsafe_decode() support
Add an intrinsic-like pseudo-function __khp_unsafe_decode() that forces
decoding of a value, independent of how the value is being used.
This can e.g. be used before:

 - handing off pointers to functions implemented in assembly
 - performing virtual-to-physical pointer translation
 - handing pointers to code that can not be recompiled to include
   HeapProtector support (e.g. EFI runtime services)

We still generate stack pins, so that we can guarantee safety if the
decoded pointer is immediately used in inline assembly or a function
call from which it can't escape. However, we naturally can't
automatically provide full protection if the decoded pointer then
escapes; this is why the name of the pseudo-function contains the word
"unsafe".