[HeapProtector] Specialized handling for single loads
The relative instrumentation overhead is worst in situations where
pointer decoding has to be invoked to compute the raw pointer for a
single, non-repeated load.

Pointer decoding is costly because it has to track which pointer the
current execution context is referencing; the reference
implementation of __khp_decode_ptr() has to:

 - handle compiler misspeculation gracefully
 - check whether the pointer is being accessed from a different
   execution context than the one it was originally allocated from
 - write the metadata pointer into an associated stack slot

But actually, we don't need any of this to perform a single load with
UAF protection: We can just perform the load first, then do the cookie
check afterwards (and on architectures that can do load-load reordering,
either a memory barrier or a single 16-byte load). Luckily a load alone
should have no overly terrible effects as long as the loaded value can't
actually be used until we know that the allocation was still live at the
time of load.

(The worst that could happen with this would be memory-mapped I/O, if it
is possible for MMIO addresses to appear in the metadata structure -
that could theoretically have side effects. An implementation should
avoid spuriously placing such values in the raw pointer field of the
metadata structure if the architecture has mappings of MMIO devices
where reads trigger significant side effects.)